requirements.txt
streamlit
tensorflow
pillow
numpy
plotly
pandas
python-dotenv
